refactor: move `actions` and `lib` outside of `app`
fix import statements to @
