+++ app/api/recipes/create/route.ts
@@ -1,4 +1,4 @@
-import { prisma } from '@/app/lib/db';
+import { prisma } from '@/lib/db';
 import { NextResponse } from 'next/server';
 
 /**
