+++ components/Button.tsx
@@ -3,7 +3,7 @@ import NextLink from 'next/link';
 
 export function Button(props: { href: string; label: string }) {
   return (
-    <MuiButton variant="text" href={props.href} LinkComponent={NextLink}>
+    <MuiButton variant="text" LinkComponent={NextLink} {...props}>
       {props.label}
     </MuiButton>
   );
